requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-dotenv==1.0.1
